+++ puma/tests/test_plot_base.py
@@ -6,7 +6,7 @@ Unit test script for the functions in plot_base.py
 
 import unittest
 
-from puma import PlotObject
+from puma import PlotBase, PlotObject
 from puma.utils import logger, set_log_level
 
 set_log_level(logger, "DEBUG")
@@ -44,3 +44,26 @@ class PlotObjectTestCase(unittest.TestCase):
         """Test wrong n ratio panels."""
         with self.assertRaises(ValueError):
             PlotObject(n_ratio_panels=5)
+
+    def test_wrong_ymin_ratio(self):
+        """Test wrong ymin_ratio."""
+        with self.assertRaises(ValueError):
+            PlotObject(n_ratio_panels=2, ymin_ratio=[0])
+
+    def test_wrong_ymax_ratio(self):
+        """Test wrong ymax_ratio."""
+        with self.assertRaises(ValueError):
+            PlotObject(n_ratio_panels=2, ymax_ratio=[0])
+
+
+class PlotBaseTestCase(unittest.TestCase):
+    """Test class for the puma.PlotBase class."""
+
+    def test_ymin_ymax_ratio(self):
+        """Test correct ymax/ymin_ratio."""
+        plot_object = PlotBase(n_ratio_panels=2, ymin_ratio=[0, 1], ymax_ratio=[1, 2])
+        plot_object.initialise_figure()
+        plot_object.set_y_lim()
+        for i in range(2):
+            ymin, ymax = plot_object.ratio_axes[i].get_ylim()
+            self.assertEqual((ymin, ymax), (i, i + 1))